+++ platformio/builder/main.py
@@ -46,8 +46,8 @@ commonvars.AddVariables(
     # library options
     ("LIB_LDF_MODE",),
     ("LIB_COMPAT_MODE",),
+    ("LIB_DEPS",),
     ("LIB_IGNORE",),
-    ("LIB_FORCE",),
     ("LIB_EXTRA_DIRS",),
 
     # board options
@@ -121,10 +121,10 @@ for var in ("BUILD_FLAGS", "SRC_BUILD_FLAGS", "SRC_FILTER", "EXTRA_SCRIPT",
         env[var] = environ.get(k)
 
 # Parse comma separated items
-for opt in ("LIB_IGNORE", "LIB_FORCE", "LIB_EXTRA_DIRS"):
+for opt in ("LIB_DEPS", "LIB_IGNORE", "LIB_EXTRA_DIRS"):
     if opt not in env:
         continue
-    env[opt] = [l.strip() for l in env[opt].split(",") if l.strip()]
+    env[opt] = [l.strip() for l in env[opt].split(", ") if l.strip()]
 
 env.Prepend(LIBSOURCE_DIRS=env.get("LIB_EXTRA_DIRS", []))
 env.LoadPioPlatform(commonvars)